Code updated to match with PHP 8+
This fatal error was appearing on opening any post after suffixing "?amp_markup=1" in the post's URL:

Fatal error: Uncaught TypeError: implode(): Argument #2 ($array) must be of type ?array, string given in .../wp-content/plugins/fb-instant-articles/vendor/facebook/facebook-instant-articles-sdk-extensions-in-php/src/Facebook/InstantArticles/AMP/AMPHeader.php:82

PHP's implode() function was supporting two arguments with it in any order. This flexibility is removed in PHP 8; and as the order used in the file is not supported in PHP 8, I changed the arguments' order according to its supported order.